containers/data_synthesizer_container/requirements.txt
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
redis==5.0.1
psycopg2-binary==2.9.7
prometheus-client==0.17.1
plotly==5.17.0
flask==2.3.3
flask-cors==4.0.0
